refactoring events event_transaction views into better file structure; adding basic events event_transaction tests; refactoring events rsvp_transaction view tests to use the same setUp method
