+++ hknweb/events/tests/views/rsvp_transactions/test_confirm_rsvp.py
@@ -1,45 +1,23 @@
 from django.test import TestCase
 
 from django.urls import reverse
-from django.contrib.auth.models import Permission, Group
-from django.contrib.contenttypes.models import ContentType
+from django.contrib.auth.models import Group
 
 from hknweb.events.models import Rsvp
 
-from hknweb.events.tests.models.utils import ModelFactory
+from hknweb.events.tests.views.rsvp_transactions.utils import setUp
 
 
-class RsvpViewTests(TestCase):
+class ConfirmRsvpViewTests(TestCase):
     def setUp(self):
-        user = ModelFactory.create_user()
-        password = "custom password"
-        user.set_password(password)
-        user.save()
-
-        content_type = ContentType.objects.get_for_model(Rsvp)
-        permission = Permission.objects.get(content_type=content_type, codename="add_rsvp")
-        user.user_permissions.add(permission)
-
-        self.client.login(username=user.username, password=password)
-
-        event_type = ModelFactory.create_event_type()
-        event = ModelFactory.create_event(
-            name="custom event name",
-            event_type=event_type,
-            created_by=user,
-            access_level=2,
-        )
+        setUp(self, ["add_rsvp"])
 
         kwargs = {
-            "id": event.id,
+            "id": self.event.id,
         }
         self.client.post(reverse("events:rsvp", kwargs=kwargs))
 
         rsvp = Rsvp.objects.first()
-
-        self.user = user
-        self.password = password
-        self.event = event
         self.rsvp = rsvp
 
     def test_not_post_returns_http404(self):
