+++ hknweb/events/views/event_transactions/__init__.py
@@ -1,0 +1,3 @@
+from hknweb.events.views.event_transactions.add_event import add_event
+from hknweb.events.views.event_transactions.update_event import EventUpdateView
+from hknweb.events.views.event_transactions.show_event import show_details
